fix: make sure the memory staging is sequential to minimize memory use

+++ crates/rho_fitting_types/src/fields/mod.rs
@@ -3,7 +3,7 @@ use ndarray::{Array4, Array5, Array6, ArrayD};
 pub const COMPONENT_COUNT: usize = 3;
 
 /// Dynamic ndarray with canonical shape `(T, Nx, Ntheta, Nr, 3, 3, 3)`.
-pub type CurrentQField = ArrayD<f64>;
+pub type CurrentQField = ArrayD<f32>;
 
 /// All particle arrays passed to mechanical deposition
 #[derive(Clone)]
@@ -17,12 +17,12 @@ pub struct ParticleFieldSet<'a> {
 
 /// Canonical raw mechanical fields. Grid axes `(T, Nx, Ntheta, Nr)` always come first.
 pub struct MechanicalFieldSet {
-    pub rho: Array4<f64>,
-    pub p: Array5<f64>,
-    pub q: Array6<f64>,
-    pub a: Array6<f64>,
-    pub hexatic_order: Array4<f64>,
-    pub j_rho: Array5<f64>,
-    pub j_p: Array6<f64>,
+    pub rho: Array4<f32>,
+    pub p: Array5<f32>,
+    pub q: Array6<f32>,
+    pub a: Array6<f32>,
+    pub hexatic_order: Array4<f32>,
+    pub j_rho: Array5<f32>,
+    pub j_p: Array6<f32>,
     pub j_q: CurrentQField,
 }
